requirements.txt
pdfplumber==0.11.4
groq
python-dotenv